adjusted variable configuration for production database

+++ models.py
@@ -23,7 +23,8 @@ def setup_db(app, database_path=database_path):
     app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
     db.app = app
     db.init_app(app)
-    db.drop_all()
+    if os.getenv("FLASK_ENV") == 'development':
+        db.drop_all()
     db.create_all()
 
 def db_rollback():
